All revenue amounts for Super Jetroid

This summary table provides revenue data for every in-app purchase and ad view for players of Super Jetroid in 2015.

all_revenue(
  size = c("small", "medium", "large", "xlarge", "preview"),
  quality = c("perfect", "faulty"),
  type = c("tibble", "data.frame", "duckdb", "csv"),
  keep = FALSE
)

Arguments

size

A keyword that allows getting different variants of the table based on the size of player base. The default "small" table has the lowest number of players/records. Increasing in size, we can also opt for the "medium", "large", or "xlarge" versions.

quality

The data quality level of the returned dataset. There are two options: (1) "perfect" provides a pristine table with no errors at all and (2) "faulty" gives you a table with a multitude of errors.

type

The table return type. By default, this is a "tibble" but a "data.frame" can instead be returned if using that keyword. If you have the duckdb package installed, you can instead obtain the table as an in-memory DuckDB database table.

keep

Should the downloaded data be stored on disk in the working directory? By default, this is FALSE. If the file is available in the next invocation then the data won't be downloaded again.

Value

A data table object, which could be a tibble (tbl_df) a data frame, or an in-memory DuckDB table (tbl_dbi). If a CSV is written then TRUE will be invisibly returned.

Examples

# Get a preview of the `all_revenue` dataset
# with the 'preview' size option
all_revenue(size = "preview")
#> # A tibble: 200 × 11
#>    player_id     sessi…¹ session_start       time                item_…² item_…³
#>    <chr>         <chr>   <dttm>              <dttm>              <chr>   <chr>  
#>  1 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 01:45:33 2015-01-01 01:56:51 iap     gold4  
#>  2 SBPFOHCVMNQI… SBPFOH… 2015-01-01 10:13:16 2015-01-01 10:17:10 ad      ad_5sec
#>  3 SBPFOHCVMNQI… SBPFOH… 2015-01-01 10:13:16 2015-01-01 10:17:46 ad      ad_5sec
#>  4 SBPFOHCVMNQI… SBPFOH… 2015-01-01 10:13:16 2015-01-01 10:19:10 iap     offer3 
#>  5 SBPFOHCVMNQI… SBPFOH… 2015-01-01 10:13:16 2015-01-01 10:22:58 ad      ad_5sec
#>  6 SBPFOHCVMNQI… SBPFOH… 2015-01-01 10:13:16 2015-01-01 10:25:52 ad      ad_5sec
#>  7 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 11:01:32 2015-01-01 11:09:26 ad      ad_15s…
#>  8 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 11:01:32 2015-01-01 11:10:14 ad      ad_15s…
#>  9 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 11:01:32 2015-01-01 11:12:50 ad      ad_sur…
#> 10 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 11:01:32 2015-01-01 11:13:38 ad      ad_sur…
#> # … with 190 more rows, 5 more variables: item_revenue <dbl>,
#> #   session_duration <dbl>, start_day <date>, acquisition <chr>, country <chr>,
#> #   and abbreviated variable names ¹​session_id, ²​item_type, ³​item_name