All player sessions for Super Jetroid

This table provides information on player sessions and summarizes the number of revenue events (ad views and IAP spends) and provides total revenue amounts (in USD) broken down by type for the session.

all_sessions(
  size = c("small", "medium", "large", "xlarge", "preview"),
  quality = c("perfect", "faulty"),
  type = c("tibble", "data.frame", "duckdb", "csv"),
  keep = FALSE
)

Arguments

size

A keyword that allows getting different variants of the table based on the size of player base. The default "small" table has the lowest number of players/records. Increasing in size, we can also opt for the "medium", "large", or "xlarge" versions.

quality

The data quality level of the returned dataset. There are two options: (1) "perfect" provides a pristine table with no errors at all and (2) "faulty" gives you a table with a multitude of errors.

type

The table return type. By default, this is a "tibble" but a "data.frame" can instead be returned if using that keyword. If you have the duckdb package installed, you can instead obtain the table as an in-memory DuckDB database table.

keep

Should the downloaded data be stored on disk in the working directory? By default, this is FALSE. If the file is available in the next invocation then the data won't be downloaded again.

Value

A data table object, which could be a tibble (tbl_df) a data frame, or an in-memory DuckDB table (tbl_dbi). If a CSV is written then TRUE will be invisibly returned.

Examples

# Get a preview of the `all_sessions` dataset
# with the 'preview' size option
all_sessions(size = "preview")
#> # A tibble: 200 × 9
#>    player_id     sessi…¹ session_start       sessi…² n_iap n_ads rev_iap rev_ads
#>    <chr>         <chr>   <dttm>                <dbl> <int> <int>   <dbl>   <dbl>
#>  1 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 01:45:33    12.8     1     0   14.0    0    
#>  2 YMKOHGVFZWLJ… YMKOHG… 2015-01-01 11:01:32    31.5     1     5   20.3    3.46 
#>  3 SBPFOHCVMNQI… SBPFOH… 2015-01-01 10:13:16    34.9     1     4   13.5    0.18 
#>  4 SBPFOHCVMNQI… SBPFOH… 2015-01-02 01:15:48    32.4     1     3    8.99   0.15 
#>  5 SBPFOHCVMNQI… SBPFOH… 2015-01-02 14:01:35    23.8     1     3   26.1    0.324
#>  6 SBPFOHCVMNQI… SBPFOH… 2015-01-03 09:54:06    24.8     1     1    4.49   0.018
#>  7 SBPFOHCVMNQI… SBPFOH… 2015-01-03 17:10:46     7.2     0     1    0      0.06 
#>  8 DLCKJEZRSIGW… DLCKJE… 2015-01-02 14:16:25     5.4     1     1    3.49   0.41 
#>  9 RNWFDZOSQPMB… RNWFDZ… 2015-01-03 07:14:57    10.1     1     2    5.49   0.42 
#> 10 RNWFDZOSQPMB… RNWFDZ… 2015-01-04 04:44:54     8       1     1    2.19   0.024
#> # … with 190 more rows, 1 more variable: rev_all <dbl>, and abbreviated
#> #   variable names ¹​session_id, ²​session_duration