Get a data frame with metadata for all datasets

The intendo_metadata() function provides a data frame containing metadata for datasets in the intendo package.

intendo_metadata(include = c("all", "perfect", "faulty"))

Arguments

include

Should both perfect and faulty datasets be included in the metadata table? This is the default ("all"), otherwise one can use the "perfect" or "faulty" keywords to get a subset of the metadata table.

Value

A data frame.

Examples

# Obtain metadata on all datasets in the package but
# only those in their 'perfect' form
intendo_metadata(include = "perfect")
#> # A tibble: 16 × 6
#>    name                basename     size   type       rows columns
#>    <chr>               <chr>        <chr>  <chr>     <int>   <int>
#>  1 all_revenue_small   all_revenue  small  perfect  195917      11
#>  2 users_daily_small   users_daily  small  perfect   44482      23
#>  3 user_summary_small  user_summary small  perfect    1231       6
#>  4 all_sessions_small  all_sessions small  perfect   61465       9
#>  5 all_revenue_medium  all_revenue  medium perfect 1026174      11
#>  6 users_daily_medium  users_daily  medium perfect  234574      23
#>  7 user_summary_medium user_summary medium perfect    6155       6
#>  8 all_sessions_medium all_sessions medium perfect  322926       9
#>  9 all_revenue_large   all_revenue  large  perfect 2093359      11
#> 10 users_daily_large   users_daily  large  perfect  478542      23
#> 11 user_summary_large  user_summary large  perfect   12311       6
#> 12 all_sessions_large  all_sessions large  perfect  658677       9
#> 13 all_revenue_xlarge  all_revenue  xlarge perfect 3739243      11
#> 14 users_daily_xlarge  users_daily  xlarge perfect  851607      23
#> 15 user_summary_xlarge user_summary xlarge perfect   20326       6
#> 16 all_sessions_xlarge all_sessions xlarge perfect 1165754       9