User summaries for Super Jetroid

This summary table provides information on each user. We get information here such as the first login/session time and some information useful for segmentation.

user_summary(
  size = c("small", "medium", "large", "xlarge", "preview"),
  quality = c("perfect", "faulty"),
  type = c("tibble", "data.frame", "duckdb", "csv"),
  keep = FALSE
)

Arguments

size

A keyword that allows getting different variants of the table based on the size of player base. The default "small" table has the lowest number of players/records. Increasing in size, we can also opt for the "medium", "large", or "xlarge" versions.

quality

The data quality level of the returned dataset. There are two options: (1) "perfect" provides a pristine table with no errors at all and (2) "faulty" gives you a table with a multitude of errors.

type

The table return type. By default, this is a "tibble" but a "data.frame" can instead be returned if using that keyword. If you have the duckdb package installed, you can instead obtain the table as an in-memory DuckDB database table.

keep

Should the downloaded data be stored on disk in the working directory? By default, this is FALSE. If the file is available in the next invocation then the data won't be downloaded again.

Value

A data table object, which could be a tibble (tbl_df) a data frame, or an in-memory DuckDB table (tbl_dbi). If a CSV is written then TRUE will be invisibly returned.

Examples

# Get a preview of the `user_summary` dataset
# with the 'preview' size option
user_summary(size = "preview")
#> # A tibble: 200 × 6
#>    player_id       first_login         start_day  country       acquis…¹ devic…²
#>    <chr>           <dttm>              <date>     <chr>         <chr>    <chr>  
#>  1 YMKOHGVFZWLJ836 2015-01-01 01:45:33 2015-01-01 United States organic  Samsun…
#>  2 SBPFOHCVMNQI568 2015-01-01 10:13:16 2015-01-01 Japan         apple    Sony E…
#>  3 DLCKJEZRSIGW561 2015-01-02 14:16:25 2015-01-02 United States organic  Sony E…
#>  4 IGOFAEVXNTUW479 2015-01-03 06:25:51 2015-01-03 India         organic  Samsun…
#>  5 RNWFDZOSQPMB164 2015-01-03 07:14:57 2015-01-03 Switzerland   google   Sony E…
#>  6 SOBKUJPREVCZ874 2015-01-03 09:15:56 2015-01-03 China         apple    iPad m…
#>  7 LKCUQHXDRWAZ134 2015-01-03 09:29:00 2015-01-03 China         facebook 2nd Ge…
#>  8 TDRXCWIOZEYM241 2015-01-03 12:58:30 2015-01-03 Denmark       other_c… 3rd Ge…
#>  9 NKBLRWVXIYJZ836 2015-01-04 00:01:14 2015-01-04 Hong Kong     apple    Sony E…
#> 10 IYOPSWCHJBRA492 2015-01-04 02:16:15 2015-01-04 South Africa  google   4th Ge…
#> # … with 190 more rows, and abbreviated variable names ¹​acquisition,
#> #   ²​device_name