Data Dictionary for user_summary

The user_summary_dd() function generates a data dictionary based on the user_summary table.

user_summary_dd(
  size = c("small", "medium", "large", "xlarge", "preview"),
  quality = c("perfect", "faulty"),
  type = c("tibble", "data.frame", "duckdb")
)

Arguments

size

A keyword that allows getting different variants of the table based on the size of player base. The default "small" table has the lowest number of players/records. Increasing in size, we can also opt for the "medium", "large", or "xlarge" versions.

quality

The data quality level of the returned dataset. There are two options: (1) "perfect" provides a pristine table with no errors at all and (2) "faulty" gives you a table with a multitude of errors.

type

The table return type. By default, this is a "tibble" but a "data.frame" can instead be returned if using that keyword. If you have the duckdb package installed, you can instead obtain the table as an in-memory DuckDB database table.

Value

A ptblank_informant object.

Examples

# Get a preview of the `user_summary` dataset
# with the 'preview' size option
user_summary_dd(size = "preview")
#> # A tibble: 200 × 6
#>    player_id      first_login         start_day  country acquisition device_name
#>    <chr>          <dttm>              <date>     <chr>   <chr>       <chr>      
#>  1 YMKOHGVFZWLJ8… 2015-01-01 01:45:33 2015-01-01 United… organic     Samsung Ga…
#>  2 SBPFOHCVMNQI5… 2015-01-01 10:13:16 2015-01-01 Japan   apple       Sony Exper…
#>  3 DLCKJEZRSIGW5… 2015-01-02 14:16:25 2015-01-02 United… organic     Sony Exper…
#>  4 IGOFAEVXNTUW4… 2015-01-03 06:25:51 2015-01-03 India   organic     Samsung Ga…
#>  5 RNWFDZOSQPMB1… 2015-01-03 07:14:57 2015-01-03 Switze… google      Sony Exper…
#>  6 SOBKUJPREVCZ8… 2015-01-03 09:15:56 2015-01-03 China   apple       iPad mini  
#>  7 LKCUQHXDRWAZ1… 2015-01-03 09:29:00 2015-01-03 China   facebook    2nd Gen iP…
#>  8 TDRXCWIOZEYM2… 2015-01-03 12:58:30 2015-01-03 Denmark other_camp… 3rd Gen iP…
#>  9 NKBLRWVXIYJZ8… 2015-01-04 00:01:14 2015-01-04 Hong K… apple       Sony Exper…
#> 10 IYOPSWCHJBRA4… 2015-01-04 02:16:15 2015-01-04 South … google      4th Gen iP…
#> # ℹ 190 more rows